changed type of sizeUpper and sizeLower to be doubles to match the entity classes

+++ frontend/search_cms/lib/features/dashboard/data/data_sources/abstract_insert_artifact_api.dart
@@ -25,8 +25,8 @@ abstract class AbstractInsertArtifactApi {
   Future<void> insertArtifact({
     required String assemblageName,
     int? porosity,
-    int? sizeUpper,
-    int? sizeLower,
+    double? sizeUpper,
+    double? sizeLower,
     String? comment,
     int? preExcavFrags,
     int? postExcavFrags,
